thanks Rangedz ;P

+++ Assembly-CSharp/SRPG/TowerRewardItem.cs
@@ -1,7 +1,7 @@
 ﻿// Decompiled with JetBrains decompiler
 // Type: SRPG.TowerRewardItem
-// Assembly: Assembly-CSharp, Version=1.2.0.0, Culture=neutral, PublicKeyToken=null
-// MVID: 9BA76916-D0BD-4DB6-A90B-FE0BCC53E511
+// Assembly: Assembly-CSharp, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null
+// MVID: FE644F5D-682F-4D6E-964D-A0DD77A288F7
 // Assembly location: C:\Users\André\Desktop\Assembly-CSharp.dll
 
 namespace SRPG
@@ -20,7 +20,7 @@ namespace SRPG
         throw new InvalidJSONException();
       this.iname = json.iname;
       this.type = (TowerRewardItem.RewardType) json.type;
-      this.num = (int) json.num;
+      this.num = json.num;
       this.visible = (int) json.visible == 1;
     }
